pyproject.toml
[project]
name = "asl-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "colorama>=0.4.6",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "loguru>=0.7.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.32.0",
]
